feat: ssr rendering aspect 함수 추가
